Add files via upload
The following code is a personal project of mine. It forecasts a given baseball player's future performance based on his previous 3 seasons of performance. This is then used to analyze and compare players to one another. To identify players that are similar to one another I have leveraged the K-Means algorithm from sklearn. 

The final product allows a user to input the name of given baseball players and view his projected statistics for the 2021 season, a list of similar players, the players 3-year average, and will see plots of key performance metrics for the previous 3 seasons